os/+arch: polybar: refactor run.sh

+++ os/+arch/.config/polybar/run.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-killall -q polybar
-while pgrep -x polybar >/dev/null; do sleep 1; done
+if pgrep polybar; then
+    killall -q polybar
+    while pgrep -x polybar >/dev/null; do sleep 1; done
+fi
 
-polybar main >~/.config/polybar/polybar.log 2>&1 &
+polybar main >>~/.config/polybar/polybar.log 2>&1 &
